src/main/java/com/localjobs/config/ApplicationConfig.java
package com.localjobs.config;

import javax.inject.Inject;

import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.ComponentScan;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.annotation.PropertySource;
import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
import org.springframework.data.mongodb.core.MongoTemplate;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.datasource.DataSourceTransactionManager;
import org.springframework.transaction.PlatformTransactionManager;
import org.springframework.transaction.annotation.EnableTransactionManagement;

@ComponentScan(basePackages = "com.localjobs")
@Configuration
@PropertySource("classpath:application.properties")
@EnableTransactionManagement
public class ApplicationConfig {
	
	@Inject
	private DatasourceConfig datasourceConfig;

	@Bean
	public PlatformTransactionManager transactionManager() {
		return new DataSourceTransactionManager(datasourceConfig.dataSource());
	}

	@Bean
	public JdbcTemplate jdbcTemplate() {
		return new JdbcTemplate(datasourceConfig.dataSource());
	}

	@Bean
	public static PropertySourcesPlaceholderConfigurer propertyPlaceHolderConfigurer() {
		return new PropertySourcesPlaceholderConfigurer();
	}

	@Bean
	public MongoTemplate mongoTemplate() throws Exception {
		MongoTemplate mongoTemplate = new MongoTemplate(datasourceConfig.mongoDbFactory());
		return mongoTemplate;
	}
}
